自定义接口返回定义

+++ yshop-common/yshop-common-core/src/main/java/com/fly4j/yshop/common/exception/GlobalExceptionHandler.java
@@ -1,9 +1,8 @@
 package com.fly4j.yshop.common.exception;
 
-import com.baomidou.mybatisplus.extension.api.IErrorCode;
-import com.baomidou.mybatisplus.extension.api.R;
-import com.baomidou.mybatisplus.extension.enums.ApiErrorCode;
-import com.baomidou.mybatisplus.extension.exceptions.ApiException;
+import com.fly4j.yshop.common.api.IErrorCode;
+import com.fly4j.yshop.common.api.R;
+import com.fly4j.yshop.common.enums.ApiErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -21,7 +20,7 @@ public class GlobalExceptionHandler {
         log.error(">> API调用异常:{}", e.getMessage());
         return R.failed(new IErrorCode() {
             @Override
-            public long getCode() {
+            public int getCode() {
                 return ApiErrorCode.FAILED.getCode();
             }
 
@@ -37,7 +36,7 @@ public class GlobalExceptionHandler {
         log.error(">> 运行时异常:{}", e.getMessage());
         return R.failed(new IErrorCode() {
             @Override
-            public long getCode() {
+            public int getCode() {
                 return ApiErrorCode.FAILED.getCode();
             }
 
@@ -53,7 +52,7 @@ public class GlobalExceptionHandler {
         log.error(">> 系统异常，请联系系统管理员,{}", e.getMessage());
         return R.failed(new IErrorCode() {
             @Override
-            public long getCode() {
+            public int getCode() {
                 return 500;
             }
